+++ packages/dso-toolkit/src/components/card/card.models.ts
@@ -3,12 +3,15 @@ import { Label } from "../label/label.models.js";
 import { Selectable } from "../selectable/selectable.models.js";
 import { Toggletip } from "../toggletip/toggletip.models.js";
 
+export type ImageShape = "square" | "rectangle";
+
 export interface Card<TemplateFnReturnType> {
   label: string;
   selectable?: Selectable<TemplateFnReturnType>;
   content: TemplateFnReturnType;
   interactions?: Array<Button | Label | Toggletip<TemplateFnReturnType>>;
   image?: string;
+  wideImage?: boolean;
   clickable?: boolean;
   dsoCardClicked?: (e: CustomEvent<DsoCardClickedEvent>) => void;
 }