+++ packages/react/src/components/card/card.react-template.tsx
@@ -9,9 +9,22 @@ export const reactCard: ComponentImplementation<Card<JSX.Element>> = {
   component: "card",
   implementation: "react",
   template: ({ iconTemplate, labelTemplate, toggletipTemplate, selectableTemplate }) =>
-    function cardTemplate({ label, selectable, content, interactions, image, clickable, dsoCardClicked }) {
+    function cardTemplate({
+      label,
+      selectable,
+      content,
+      interactions,
+      image,
+      rectangularImage,
+      clickable,
+      dsoCardClicked,
+    }) {
       return (
-        <DsoCard clickable={clickable} onDsoCardClicked={(e: CustomEvent<DsoCardClickedEvent>) => dsoCardClicked?.(e)}>
+        <DsoCard
+          clickable={clickable}
+          rectangularImage={rectangularImage}
+          onDsoCardClicked={(e: CustomEvent<DsoCardClickedEvent>) => dsoCardClicked?.(e)}
+        >
           {selectable && selectableTemplate(selectable)}
           {image && <img slot="image" src={image} />}
           {clickable ? (